src/main/java/ejercicios/Division.java

package ejercicios;

import static ejercicios.SetDeTenis.evaluar;
import java.util.Scanner;

/**
 *
 * @author danielsanchez
 */
public class Division {
    public static String evaluar(int dividendo, int divisor) {
        int cociente = 0;
        int residuo = 0;
        String exacta ="";
        residuo = dividendo%divisor;
        cociente = dividendo/divisor;
        if (residuo == 0){
            exacta = "es";
        }else{
            exacta = "no es";
        }
        String respuesta = "La división " + exacta + " exacta. \n"
                + "Cociente: " + cociente + "\n"
                + "Residuo: " + residuo;
        return respuesta;
    }
    
    public static void main(String[] args) {
        Scanner lector = new Scanner(System.in);
        System.out.print("Dividendo:");
        int dividendo = lector.nextInt();
        System.out.print("Divisor:");
        int divisor = lector.nextInt();

        String respuesta = evaluar(dividendo, divisor);
        System.out.println(respuesta);
    }
}
